+++ ste/sender-azureFile.go
@@ -34,14 +34,27 @@ import (
 	"github.com/Azure/azure-storage-azcopy/common"
 )
 
+type URLHolder interface {
+	URL() url.URL
+	String() string
+}
+
+// TODO: review the decision to have this implement both interfaces
+// azureFileSenderBase implements both IFolderSender and (most of) IFileSender.
+// Why implement both interfaces in the one type, even though they are largely unrelated? Because it
+// makes functions like newAzureFilesUploader easier to reason about, since they always return the same type.
+// It may also make it easier to describe what's needed when supporting an new backend - e.g. "to send to a new back end
+// you need a sender that implements IFileSender and, if the back end is folder aware, it should also implement IFolderSender"
+// (The alternative would be to have the likes of newAzureFilesUploader call sip.EntityType and return a different type
+// if the entity type is folder).
 type azureFileSenderBase struct {
-	jptm      IJobPartTransferMgr
-	fileURL   azfile.FileURL
-	chunkSize uint32
-	numChunks uint32
-	pipeline  pipeline.Pipeline
-	pacer     pacer
-	ctx       context.Context
+	jptm         IJobPartTransferMgr
+	fileOrDirURL URLHolder
+	chunkSize    uint32
+	numChunks    uint32
+	pipeline     pipeline.Pipeline
+	pacer        pacer
+	ctx          context.Context
 	// Headers and other info that we will apply to the destination
 	// object. For S2S, these come from the source service.
 	// When sending local data, they are computed based on
@@ -54,7 +67,7 @@ func newAzureFileSenderBase(jptm IJobPartTransferMgr, destination string, p pipe
 
 	info := jptm.Info()
 
-	// compute chunk size
+	// compute chunk size (irrelevant but harmless for folders)
 	// If the given chunk Size for the Job is greater than maximum file chunk size i.e 4 MB
 	// then chunk size will be 4 MB.
 	chunkSize := info.BlockSize
@@ -66,7 +79,7 @@ func newAzureFileSenderBase(jptm IJobPartTransferMgr, destination string, p pipe
 		}
 	}
 
-	// compute num chunks
+	// compute num chunks (irrelevant but harmless for folders)
 	numChunks := getNumChunks(info.SourceSize, chunkSize)
 
 	// make sure URL is parsable
@@ -78,14 +91,22 @@ func newAzureFileSenderBase(jptm IJobPartTransferMgr, destination string, p pipe
 	// due to the REST parity feature added in 2019-02-02, the File APIs are no longer backward compatible
 	// so we must use the latest SDK version to stay safe
 	ctx := context.WithValue(jptm.Context(), ServiceAPIVersionOverride, azfile.ServiceVersion)
+
 	props, err := sip.Properties()
 	if err != nil {
 		return nil, err
 	}
 
+	var h URLHolder
+	if info.IsFolderPropertiesTransfer() {
+		h = azfile.NewDirectoryURL(*destURL, p)
+	} else {
+		h = azfile.NewFileURL(*destURL, p)
+	}
+
 	return &azureFileSenderBase{
 		jptm:            jptm,
-		fileURL:         azfile.NewFileURL(*destURL, p),
+		fileOrDirURL:    h,
 		chunkSize:       chunkSize,
 		numChunks:       numChunks,
 		pipeline:        p,
@@ -94,6 +115,22 @@ func newAzureFileSenderBase(jptm IJobPartTransferMgr, destination string, p pipe
 		headersToApply:  props.SrcHTTPHeaders.ToAzFileHTTPHeaders(),
 		metadataToApply: props.SrcMetadata.ToAzFileMetadata(),
 	}, nil
+}
+
+func (u *azureFileSenderBase) fileURL() azfile.FileURL {
+	return u.fileOrDirURL.(azfile.FileURL)
+}
+
+func (u *azureFileSenderBase) dirURL() azfile.DirectoryURL {
+	return u.fileOrDirURL.(azfile.DirectoryURL)
+}
+
+func (u *azureFileSenderBase) SendableEntityType() common.EntityType {
+	if _, ok := u.fileOrDirURL.(azfile.DirectoryURL); ok {
+		return common.EEntityType.Folder()
+	} else {
+		return common.EEntityType.File()
+	}
 }
 
 func (u *azureFileSenderBase) ChunkSize() uint32 {
@@ -105,7 +142,7 @@ func (u *azureFileSenderBase) NumChunks() uint32 {
 }
 
 func (u *azureFileSenderBase) RemoteFileExists() (bool, error) {
-	return remoteObjectExists(u.fileURL.GetProperties(u.ctx))
+	return remoteObjectExists(u.fileURL().GetProperties(u.ctx))
 }
 
 func (u *azureFileSenderBase) Prologue(state common.PrologueState) (destinationModified bool) {
@@ -115,7 +152,7 @@ func (u *azureFileSenderBase) Prologue(state common.PrologueState) (destinationM
 	destinationModified = true
 
 	// Create the parent directories of the file. Note share must be existed, as the files are listed from share or directory.
-	err := AzureFileParentDirCreator{}.CreateParentDirToRoot(u.ctx, u.fileURL, u.pipeline)
+	err := AzureFileParentDirCreator{}.CreateParentDirToRoot(u.ctx, u.fileURL(), u.pipeline)
 	if err != nil {
 		jptm.FailActiveUpload("Creating parent directory", err)
 		return
@@ -128,7 +165,7 @@ func (u *azureFileSenderBase) Prologue(state common.PrologueState) (destinationM
 	}
 
 	// Create Azure file with the source size
-	_, err = u.fileURL.Create(u.ctx, info.SourceSize, u.headersToApply, u.metadataToApply)
+	_, err = u.fileURL().Create(u.ctx, info.SourceSize, u.headersToApply, u.metadataToApply)
 	if err != nil {
 		jptm.FailActiveUpload("Creating file", err)
 		return
@@ -147,15 +184,15 @@ func (u *azureFileSenderBase) Cleanup() {
 		// contents will be at an unknown stage of partial completeness
 		deletionContext, cancelFn := context.WithTimeout(context.Background(), 2*time.Minute)
 		defer cancelFn()
-		_, err := u.fileURL.Delete(deletionContext)
+		_, err := u.fileURL().Delete(deletionContext)
 		if err != nil {
-			jptm.Log(pipeline.LogError, fmt.Sprintf("error deleting the (incomplete) file %s. Failed with error %s", u.fileURL.String(), err.Error()))
+			jptm.Log(pipeline.LogError, fmt.Sprintf("error deleting the (incomplete) file %s. Failed with error %s", u.fileOrDirURL.String(), err.Error()))
 		}
 	}
 }
 
 func (u *azureFileSenderBase) GetDestinationLength() (int64, error) {
-	prop, err := u.fileURL.GetProperties(u.ctx)
+	prop, err := u.fileURL().GetProperties(u.ctx)
 
 	if err != nil {
 		return -1, err
@@ -164,13 +201,26 @@ func (u *azureFileSenderBase) GetDestinationLength() (int64, error) {
 	return prop.ContentLength(), nil
 }
 
+func (u *azureFileSenderBase) RemoteFolderExists() (bool, error) {
+	return remoteObjectExists(u.dirURL().GetProperties(u.ctx))
+}
+
+func (u *azureFileSenderBase) EnsureFolderExists() error {
+	return AzureFileParentDirCreator{}.CreateDirToRoot(u.ctx, u.dirURL(), u.pipeline)
+}
+
+func (u *azureFileSenderBase) SetFolderProperties() error {
+	// TODO
+	return nil
+}
+
 // namespace for functions related to creating parent directories in Azure File
 // to avoid free floating global funcs
 type AzureFileParentDirCreator struct{}
 
 // getParentDirectoryURL gets parent directory URL of an Azure FileURL.
-func (AzureFileParentDirCreator) getParentDirectoryURL(fileURL azfile.FileURL, p pipeline.Pipeline) azfile.DirectoryURL {
-	u := fileURL.URL()
+func (AzureFileParentDirCreator) getParentDirectoryURL(uh URLHolder, p pipeline.Pipeline) azfile.DirectoryURL {
+	u := uh.URL()
 	u.Path = u.Path[:strings.LastIndex(u.Path, "/")]
 	return azfile.NewDirectoryURL(u, p)
 }
@@ -201,8 +251,12 @@ func (AzureFileParentDirCreator) splitWithoutToken(str string, token rune) []str
 // CreateParentDirToRoot creates parent directories of the Azure file if file's parent directory doesn't exist.
 func (d AzureFileParentDirCreator) CreateParentDirToRoot(ctx context.Context, fileURL azfile.FileURL, p pipeline.Pipeline) error {
 	dirURL := d.getParentDirectoryURL(fileURL, p)
+	return d.CreateDirToRoot(ctx, dirURL, p)
+}
+
+// CreateDirToRoot Creates the dir (and parents as necessary) if it does not exist
+func (d AzureFileParentDirCreator) CreateDirToRoot(ctx context.Context, dirURL azfile.DirectoryURL, p pipeline.Pipeline) error {
 	dirURLExtension := common.FileURLPartsExtension{FileURLParts: azfile.NewFileURLParts(dirURL.URL())}
-	// Check whether parent dir of the file exists.
 	if _, err := dirURL.GetProperties(ctx); err != nil {
 		if stgErr, stgErrOk := err.(azfile.StorageError); stgErrOk && stgErr.Response() != nil &&
 			stgErr.Response().StatusCode == http.StatusNotFound { // At least need read and write permisson for destination