+++ ste/xfer-anyToRemote-file.go
@@ -36,10 +36,20 @@ import (
 // This sync.Once is present to ensure we output information about a S2S access tier preservation failure to stdout once
 var s2sAccessTierFailureLogStdout sync.Once
 
-// anyToRemote handles all kinds of sender operations - both uploads from local files, and S2S copies
+// xfer.go requires just a single xfer function for the whole job.
+// This routine serves that role for uploads and S2S copies, and redirects for each transfer to a file or folder implementation
 func anyToRemote(jptm IJobPartTransferMgr, p pipeline.Pipeline, pacer pacer, senderFactory senderFactory, sipf sourceInfoProviderFactory) {
-
 	info := jptm.Info()
+	if info.IsFolderPropertiesTransfer() {
+		anyToRemote_folder(jptm, info, p, pacer, senderFactory, sipf)
+	} else {
+		anyToRemote_file(jptm, info, p, pacer, senderFactory, sipf)
+	}
+}
+
+// anyToRemote_file handles all kinds of sender operations for files - both uploads from local files, and S2S copies
+func anyToRemote_file(jptm IJobPartTransferMgr, info TransferInfo, p pipeline.Pipeline, pacer pacer, senderFactory senderFactory, sipf sourceInfoProviderFactory) {
+
 	srcSize := info.SourceSize
 
 	// step 1. perform initial checks
@@ -56,14 +66,19 @@ func anyToRemote(jptm IJobPartTransferMgr, p pipeline.Pipeline, pacer pacer, sen
 		jptm.ReportTransferDone()
 		return
 	}
+	if srcInfoProvider.EntityType() != common.EEntityType.File() {
+		panic("configuration error. Source Info Provider does not have File entity type")
+	}
 
-	s, err := senderFactory(jptm, info.Destination, p, pacer, srcInfoProvider)
+	sBase, err := senderFactory(jptm, info.Destination, p, pacer, srcInfoProvider)
 	if err != nil {
 		jptm.LogSendError(info.Source, info.Destination, err.Error(), 0)
 		jptm.SetStatus(common.ETransferStatus.Failed())
 		jptm.ReportTransferDone()
 		return
 	}
+	s := sBase.(IFileSender)
+
 	// step 2b. Read chunk size and count from the sender (since it may have applied its own defaults and/or calculations to produce these values
 	numChunks := s.NumChunks()
 	if jptm.ShouldLog(pipeline.LogInfo) {
@@ -99,7 +114,7 @@ func anyToRemote(jptm IJobPartTransferMgr, p pipeline.Pipeline, pacer pacer, sen
 			if !shouldOverwrite {
 				// logging as Warning so that it turns up even in compact logs, and because previously we use Error here
 				jptm.LogAtLevelForCurrentTransfer(pipeline.LogWarning, "File already exists, so will be skipped")
-				jptm.SetStatus(common.ETransferStatus.SkippedFileAlreadyExists())
+				jptm.SetStatus(common.ETransferStatus.SkippedEntityAlreadyExists())
 				jptm.ReportTransferDone()
 				return
 			}
@@ -121,12 +136,12 @@ func anyToRemote(jptm IJobPartTransferMgr, p pipeline.Pipeline, pacer pacer, sen
 		defer srcFile.Close() // we read all the chunks in this routine, so can close the file at the end
 	}
 
-	// Do LMT verfication before transfer, when:
+	// We always to LMT verification after the transfer. Also do it here, before transfer, when:
 	// 1) Source is local, so get source file's LMT is free.
 	// 2) Source is remote, i.e. S2S copy case. And source's size is larger than one chunk. So verification can possibly save transfer's cost.
 	if copier, isS2SCopier := s.(s2sCopier); srcInfoProvider.IsLocal() ||
 		(isS2SCopier && info.S2SSourceChangeValidation && srcSize > int64(copier.ChunkSize())) {
-		lmt, err := srcInfoProvider.GetLastModifiedTime()
+		lmt, err := srcInfoProvider.GetFileLastModifiedTime()
 		if err != nil {
 			jptm.LogSendError(info.Source, info.Destination, "Couldn't get source's last modified time-"+err.Error(), 0)
 			jptm.SetStatus(common.ETransferStatus.Failed())
@@ -183,7 +198,7 @@ var jobCancelledLocalPrefetchErr = errors.New("job was cancelled; Pre-fetching s
 // is harmless (and a good thing, to avoid excessive RAM usage).
 // To take advantage of the good sequential read performance provided by many file systems,
 // and to be able to compute an MD5 hash for the file, we work sequentially through the file here.
-func scheduleSendChunks(jptm IJobPartTransferMgr, srcPath string, srcFile common.CloseableReaderAt, srcSize int64, s ISenderBase, sourceFileFactory common.ChunkReaderSourceFactory, srcInfoProvider ISourceInfoProvider) {
+func scheduleSendChunks(jptm IJobPartTransferMgr, srcPath string, srcFile common.CloseableReaderAt, srcSize int64, s IFileSender, sourceFileFactory common.ChunkReaderSourceFactory, srcInfoProvider ISourceInfoProvider) {
 	// For generic send
 	chunkSize := s.ChunkSize()
 	numChunks := s.NumChunks()
@@ -304,7 +319,7 @@ func isDummyChunkInEmptyFile(startIndex int64, fileSize int64) bool {
 }
 
 // Complete epilogue. Handles both success and failure.
-func epilogueWithCleanupSendToRemote(jptm IJobPartTransferMgr, s ISenderBase, sip ISourceInfoProvider) {
+func epilogueWithCleanupSendToRemote(jptm IJobPartTransferMgr, s IFileSender, sip ISourceInfoProvider) {
 	info := jptm.Info()
 	// allow our usual state tracking mechanism to keep count of how many epilogues are running at any given instant, for perf diagnostics
 	pseudoId := common.NewPseudoChunkIDForWholeFile(info.Source)
@@ -314,7 +329,7 @@ func epilogueWithCleanupSendToRemote(jptm IJobPartTransferMgr, s ISenderBase, si
 	if jptm.IsLive() {
 		if _, isS2SCopier := s.(s2sCopier); sip.IsLocal() || (isS2SCopier && info.S2SSourceChangeValidation) {
 			// Check the source to see if it was changed during transfer. If it was, mark the transfer as failed.
-			lmt, err := sip.GetLastModifiedTime()
+			lmt, err := sip.GetFileLastModifiedTime()
 			if err != nil {
 				jptm.FailActiveSend("epilogueWithCleanupSendToRemote", err)
 			}
@@ -346,10 +361,14 @@ func epilogueWithCleanupSendToRemote(jptm IJobPartTransferMgr, s ISenderBase, si
 
 	jptm.UnlockDestination()
 
+	commonSenderCompletion(jptm, s, info)
+}
+
+// commonSenderCompletion is used for both files and folders
+func commonSenderCompletion(jptm IJobPartTransferMgr, s ISenderBase, info TransferInfo) {
 	if jptm.TransferStatusIgnoringCancellation() == 0 {
 		panic("think we're finished but status is notStarted")
 	}
-
 	// note that we do not really know whether the context was canceled because of an error, or because the user asked for it
 	// if was an intentional cancel, the status is still "in progress", so we are still counting it as pending
 	// we leave these transfer status alone
@@ -367,10 +386,10 @@ func epilogueWithCleanupSendToRemote(jptm IJobPartTransferMgr, s ISenderBase, si
 		// Final logging
 		if jptm.ShouldLog(pipeline.LogInfo) { // TODO: question: can we remove these ShouldLogs?  Aren't they inside Log?
 			if _, ok := s.(s2sCopier); ok {
-				jptm.Log(pipeline.LogInfo, fmt.Sprintf("COPYSUCCESSFUL: %s", strings.Split(info.Destination, "?")[0]))
+				jptm.Log(pipeline.LogInfo, fmt.Sprintf("COPYSUCCESSFUL: %s%s", info.entityTypeLogIndicator(), strings.Split(info.Destination, "?")[0]))
 			} else if _, ok := s.(uploader); ok {
 				// Output relative path of file, includes file name.
-				jptm.Log(pipeline.LogInfo, fmt.Sprintf("UPLOADSUCCESSFUL: %s", strings.Split(info.Destination, "?")[0]))
+				jptm.Log(pipeline.LogInfo, fmt.Sprintf("UPLOADSUCCESSFUL: %s%s", info.entityTypeLogIndicator(), strings.Split(info.Destination, "?")[0]))
 			} else {
 				panic("invalid state: epilogueWithCleanupSendToRemote should be used by COPY and UPLOAD")
 			}
@@ -383,7 +402,6 @@ func epilogueWithCleanupSendToRemote(jptm IJobPartTransferMgr, s ISenderBase, si
 			jptm.Log(pipeline.LogDebug, "Finalizing Transfer Cancellation/Failure")
 		}
 	}
-
 	// successful or unsuccessful, it's definitely over
 	jptm.ReportTransferDone()
 }